\name{hmgm-package}
\alias{hmgm-package}
\alias{hmgm-package}
\docType{package}

\title{
    High-dimensional mixed graphical models estimation
}

\description{
     A package for high-dimensional mixed graphical models estimation.
}


\details{
\tabular{ll}{
Package: \tab  hmgm\cr
Type: \tab Package\cr
Version: \tab 0.3.0\cr
Date: \tab 2019-11-29\cr
License: \tab GPL (>= 2)\cr
}
The major function \code{hmgm} provides weighted group lasso framework for high-dimensional mixed data graph estimation.

Another function \code{pargroup} identify all regions where groups intersect, make all variables in each overlapping region into a new group.
}

\author{
Mingyu Qi, Tianxi Li
}

\references{
Jie Cheng, Tianxi Li, Elizaveta Levina, and Ji Zhu.(2017)
 \emph{ High-dimensional Mixed Graphical Models. Journal of Computational and Graphical Statistics 26.2 (2017): 367-378},\url{https://arxiv.org/pdf/1304.2810.pdf}\cr
Simon, N., Friedman, J., Hastie,T., Tibshirani, R.(2011)
  \emph{Regularization Paths for Cox's ProportionalHazards Model via Coordinate Descent, Journal of Statistical Software, Vol.39(5) 1-13},\url{https://www.jstatsoft.org/v39/i05/}\cr
Meinshausen, N. and Buhlmann, P. (2006)
  \emph{High dimensional graphs and variable selection with the lasso, Annals of Statistics, 34, 1436–1462.},
  \url{https://arxiv.org/pdf/math/0608017.pdf}\cr
Ravikumar, P., Wainwright, M., and Lafferty, J.(2010)
\emph{ High-dimensionalIsing model selection using l1-regularized logistic regression,Annals of
Statistics, 38, 1287–1319.},\url{https://arxiv.org/pdf/1010.0311.pdf} \cr
Liu, H., Han, F., Yuan, M., Lafferty, J., and Wasserman, L.(2012)
\emph{ High dimensional semiparametric Gaussian copula graphical models, Annals
of Statistics, 40, 2293–2326.},
\url{https://arxiv.org/pdf/1202.2169.pdf}\cr
Zhao, P., Rocha, G., and Yu, B.(2009)
\emph{ The composite absolute penalties family for grouped and hierarchical variable selection, The Annals of
Statistics, 3468–3497.},
\url{https://arxiv.org/pdf/0909.0411.pdf}
}

\keyword{package}
